+++ R/loadData.R
@@ -1,5 +1,8 @@
 #' Helper function to load package data
 #' @importFrom utils data
+#' @examples
+#' testData <- loadData("DMAP_exprs")
+#' @return a spaceRATScaffolds data object
 #' @export
 loadData <- function(name){
     utils::data(