+++ tests/testthat/test-getAllScaffolds.R
@@ -1,0 +1,4 @@
+test_that("getAllScaffolds works", {
+  allScaffolds <- getAllScaffolds()
+  expect_s3_class(allScaffolds, "data.frame")
+})